#20 feat: refactored Popup to get child component as `children `
